fix: :bug: Fix bug that prevents breakpoints to be hit inside function definitions (#7)
Custom gate definitions are defined with start position at `gate...` and end position at their closing `}`. Therefore, when putting a breakpoint at any location within those bounds, it was always attached to the function. Now, if a breakpoint would be attached to a function, we first check if it would fit better to one of the gates inside it.