feat: :sparkles: Data dependency diagnosis can now step out of the current function call, even if the diagnosis was started in the same scope
Previously, data dependency analysis was only able to step outside of functions that were entered during the analysis (i.e. where the analysis knows from where the function was entered). Now, the parameter `includeCallers` can be passed to allow the analysis to step out. In this case, all possible callers will be included as dependencies.